Day 17
I've written this code along with my teacher Angela. This is quite nice to move forward and finally get some basic knowledge about classes. This is new for me and right now I've passed a line when my knowledge about programming stopped a long time ago. So let's get to the OOP :D